Prevent size_t conflicting with unsigned in some platforms
git-svn-id: https://svn.smr.arpa.emr.it/dballe/dballe/trunk@4286 5c06e07b-eeef-0310-a3fe-9f68d50baefc

+++ dballe/core/csv.cc
@@ -380,7 +380,7 @@ void CSVWriter::add_value_raw(const std::string& str)
     row.append(str);
 }
 
-void CSVWriter::add_value(size_t val)
+void CSVWriter::add_value(uint64_t val)
 {
     FormatInt fmt(val);
     add_value_raw(fmt.c_str());
